Crate eyeliner

A CSS inliner for making emails.

Modules

traits
All of the traits implemented by Eyeliner.

Structs

Eyeliner
Data and methods related to modifying HTML with CSS.

HashableNodeRef
Wraps nodes from Kuchiki to make them hashable.

Options
Options for ways to modify the HTML document using CSS.

OptionsRequired
The required options to inline HTML and CSS. Use the other Options, it is has nice defaults.

Rules
Data collected from the CSS stylesheet.

Settings
Settings referenced by features enabled through Options.

SettingsRequired
The required settings to inline HTML and CSS. Use the other Settings, it is has nice defaults.

Functions

inline
Returns a string of HTML with CSS inlined.