Struct eyeliner::Options

pub struct Options<'a> {
    pub apply_table_element_attributes: Option<bool>,
    pub apply_height_attributes: Option<bool>,
    pub apply_style_tags: Option<bool>,
    pub apply_width_attributes: Option<bool>,
    pub insert_preserved_css: Option<Vec<&'a str>>,
    pub preserve_font_faces: Option<bool>,
    pub preserve_important: Option<bool>,
    pub preserve_media_queries: Option<bool>,
    pub remove_style_tags: Option<bool>,
}

Options for ways to modify the HTML document using CSS.

Fields
apply_table_element_attributes: Option<bool>

Whether attributes specified in Settings.style_to_attribute get applied to table elements in Settings.table_elements. Defaults to true.

apply_height_attributes: Option<bool>

Whether to use any CSS pixel heights to create height attributes on elements set in Settings.height_elements. Defaults to true.

apply_style_tags: Option<bool>

Whether to inline CSS in <style /> tags in the HTML document. Defaults to true.

apply_width_attributes: Option<bool>

Whether to use any CSS pixel widths to create width attributes on elements set in Settings.width_elements. Defaults to true.

insert_preserved_css: Option<Vec<&'a str>>

List of elements to try to inline preserved @media and @font-face CSS rules into. Give an empty list to prevent inlining preserved CSS. Defaults to ["head", "body", "html"].

preserve_font_faces: Option<bool>

Whether to preserve @font-face CSS rules. Defaults to true.

preserve_important: Option<bool>

Whether to preserve !important in CSS rules. Defaults to false.

preserve_media_queries: Option<bool>

Whether to preserve @media CSS rules. Defaults to true.

remove_style_tags: Option<bool>

Whether to remove <style /> tags, after they have optionally had their CSS extracted. Defaults to true.