Struct eyeliner::Settings

pub struct Settings<'a> {
    pub width_elements: Option<Vec<&'a str>>,
    pub height_elements: Option<Vec<&'a str>>,
    pub style_to_attribute: Option<HashMap<&'a str, &'a str>>,
    pub table_elements: Option<Vec<&'a str>>,
    pub non_visual_elements: Option<Vec<&'a str>>,
    pub excluded_properties: Option<Vec<&'a str>>,
}

Settings referenced by features enabled through Options.

Fields
width_elements: Option<Vec<&'a str>>

List of HTML elements that can receive width attributes.

Defaults to:

vec![
    "table",
    "td",
    "img",
];

height_elements: Option<Vec<&'a str>>

List of HTML elements that can receive height attributes.

Defaults to:

vec![
    "table",
    "td",
    "img",
];

style_to_attribute: Option<HashMap<&'a str, &'a str>>

Map of style property names to their respective attribute names.

Defaults to:

hashmap!{
    "background-color" => "bgcolor",
    "background-image" => "background",
    "text-align" => "align",
    "vertical-align" => "valign",
};

table_elements: Option<Vec<&'a str>>

List of table HTML elements that can receive attributes defined in Settings.style_to_attribute.

Defaults to:

vec![
    "table",
    "td",
    "th",
    "tr",
    "td",
    "caption",
    "colgroup",
    "col",
    "thead",
    "tbody",
    "tfoot",
];

non_visual_elements: Option<Vec<&'a str>>

List of elements that will not have styles inlined because they are not intended to render.

Defaults to:

vec![
    "head",
    "title",
    "base",
    "link",
    "style",
    "meta",
    "script",
    "noscript",
];

excluded_properties: Option<Vec<&'a str>>

List of CSS style properties that will not be inlined.

Defaults to:

Vec::<&str>::new();